create Listings component. Move code from Carousel to Listings. Carousel now conditionally renders the Modal/Listings to change the view
